+++ libfive/include/libfive/eval/eval_xtree.hpp
@@ -18,10 +18,10 @@ class XTreeEvaluator
 {
 public:
     XTreeEvaluator(const Tree t)
-        : XTreeEvaluator(t, std::map<Tree::Id, float>())
+        : XTreeEvaluator(t, std::map<Tree::Id, double>())
     { /* Nothing to do here */ }
 
-    XTreeEvaluator(const Tree t, const std::map<Tree::Id, float>& vars)
+    XTreeEvaluator(const Tree t, const std::map<Tree::Id, double>& vars)
         : deck(new Deck(t)), array(deck, vars), interval(deck, vars),
           feature(deck, vars), deriv(deck, vars)
     { /* Nothing to do here */ }
@@ -29,7 +29,7 @@ public:
     /*
      *  Updates variable values, return true if changed
      */
-    bool updateVars(const std::map<Kernel::Tree::Id, float>& vars)
+    bool updateVars(const std::map<Kernel::Tree::Id, double>& vars)
     {
         bool changed = false;
         for (auto& v : vars)
