+++ libfive/include/libfive/eval/interval.hpp
@@ -13,11 +13,11 @@ You can obtain one at http://mozilla.org/MPL/2.0/.
 namespace Kernel {
 namespace Interval {
 
-typedef boost::numeric::interval<float,
+typedef boost::numeric::interval<double,
     boost::numeric::interval_lib::policies<
         boost::numeric::interval_lib::save_state<
-            boost::numeric::interval_lib::rounded_transc_std<float>>,
-        boost::numeric::interval_lib::checking_base<float>>> I;
+            boost::numeric::interval_lib::rounded_transc_std<double>>,
+        boost::numeric::interval_lib::checking_base<double>>> I;
 
 enum State { EMPTY, FILLED, AMBIGUOUS, UNKNOWN };
 
@@ -72,7 +72,7 @@ inline Kernel::Interval::I atan2(const Kernel::Interval::I& y,
         }
         else
         {   // Branch cut
-            return Kernel::Interval::I(-float(M_PI), float(M_PI));
+            return Kernel::Interval::I(-double(M_PI), double(M_PI));
         }
     }
     else
@@ -91,7 +91,7 @@ inline Kernel::Interval::I atan2(const Kernel::Interval::I& y,
         else
         {
             // Contains the origin
-            return Kernel::Interval::I(-float(M_PI), float(M_PI));
+            return Kernel::Interval::I(-double(M_PI), double(M_PI));
         }
     }
 }
