+++ libfive/include/libfive/render/brep/worker_pool.hpp
@@ -53,7 +53,7 @@ protected:
         boost::lockfree::stack<Task, boost::lockfree::fixed_sized<true>>;
 
     static void run(XTreeEvaluator* eval, LockFreeStack& tasks,
-                    const float max_err, std::atomic_bool& done,
+                    const double max_err, std::atomic_bool& done,
                     std::atomic_bool& cancel, Root<T>& root,
                     std::mutex& root_lock, ProgressWatcher* progress);
 };
